+++ CHANGELOG.md
@@ -11,11 +11,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.5.0-dev.1]
+## [1.5.0-dev.2]
 
 ### Fixed
 
 - Accessory could not be recovered from NO_RESPONSE using single Characteristic
+- Make unsupported Characteristic error more
+  descriptive [#456](https://github.com/NRCHKB/node-red-contrib-homekit-bridged/issues/456)
 
 ### Changed